+++ samples/SnapshotConsoleSample/SnapshotConsoleSample.c
@@ -203,8 +203,6 @@ void print_timestamp(dxf_long_t timestamp) {
 	wprintf(L"%ls", timefmt);
 }
 
-/* -------------------------------------------------------------------------- */
-
 void process_last_error() {
 	int error_code = dx_ec_success;
 	dxf_const_string_t error_descr = NULL;
@@ -227,8 +225,6 @@ void process_last_error() {
 
 	wprintf(L"An error occurred but the error subsystem failed to initialize\n");
 }
-
-/* -------------------------------------------------------------------------- */
 
 dxf_string_t ansi_to_unicode(const char *ansi_str) {
 #ifdef _WIN32
@@ -257,8 +253,6 @@ dxf_string_t ansi_to_unicode(const char *ansi_str) {
 #endif /* _WIN32 */
 }
 
-/* -------------------------------------------------------------------------- */
-
 void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 	size_t i;
 	size_t records_count = snapshot_data->records_count;
@@ -276,7 +270,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < records_count; ++i) {
 			dxf_order_t order = order_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -293,7 +287,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < snapshot_data->records_count; ++i) {
 			dxf_candle_t candle = candle_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -311,7 +305,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < records_count; ++i) {
 			dxf_order_t order = order_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -330,7 +324,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < snapshot_data->records_count; ++i) {
 			dxf_time_and_sale_t tns = time_and_sale_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -347,7 +341,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < snapshot_data->records_count; ++i) {
 			dxf_greeks_t grks = greeks_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -364,7 +358,7 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		for (i = 0; i < snapshot_data->records_count; ++i) {
 			dxf_series_t srs = series_records[i];
 
-			if (records_print_limit > 0 && i >= records_print_limit) {
+			if (records_print_limit > 0 && i >= (size_t)records_print_limit) {
 				wprintf(L"   { ... %zu records left ...}\n", records_count - i);
 				break;
 			}
@@ -379,8 +373,6 @@ void listener(const dxf_snapshot_data_ptr_t snapshot_data, void *user_data) {
 		}
 	}
 }
-
-/* -------------------------------------------------------------------------- */
 
 int atoi2(char *str, int *result) {
 	if (str == NULL || str[0] == '\0' || result == NULL) {
